Revert "Fix action discord_message_id type (#740)"
This reverts commit 80e016c4672a40991d385c9da8b6bfc54b279909.

+++ lib/teiserver/moderation/schemas/action.ex
@@ -14,7 +14,7 @@ defmodule Teiserver.Moderation.Action do
 
     field :hidden, :boolean, default: false
 
-    field :discord_message_id, :bigint
+    field :discord_message_id, :integer
 
     belongs_to :report_group, Teiserver.Moderation.ReportGroup
